+++ src/main/java/duke/commands/AddCommand.java
@@ -1,9 +1,9 @@
 package duke.commands;
 
 import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 import duke.tasks.Task;
 import duke.tasks.TaskType;
 
@@ -41,7 +41,7 @@ public class AddCommand extends Command {
     }
 
     @Override
-    public String runNew(TaskList taskList, Storage storage, Ui ui) throws DukeException {
+    public String runGUI(TaskList taskList, Storage storage, Ui ui) throws DukeException {
         Task newTask;
         if (taskType == TaskType.TODO) {
             newTask = taskList.addTodo(attributes);
