+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,9 +1,9 @@
 package duke.commands;
 
 import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 import duke.tasks.Task;
 
 /**
@@ -29,7 +29,7 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    public String runNew(TaskList taskList, Storage storage, Ui ui) throws DukeException {
+    public String runGUI(TaskList taskList, Storage storage, Ui ui) throws DukeException {
         Task deletedTask = taskList.deleteTask(Integer.parseInt(attributes));
         storage.storeList(taskList.getList());
         return ui.writeDelete(deletedTask, taskList.getSize());
