+++ src/main/java/duke/commands/DoneCommand.java
@@ -1,9 +1,9 @@
 package duke.commands;
 
 import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 import duke.tasks.Task;
 
 /**
@@ -29,7 +29,7 @@ public class DoneCommand extends Command {
     }
 
     @Override
-    public String runNew(TaskList taskList, Storage storage, Ui ui) throws DukeException {
+    public String runGUI(TaskList taskList, Storage storage, Ui ui) throws DukeException {
         Task markedTask = taskList.markDone(Integer.parseInt(attributes));
         storage.storeList(taskList.getList());
         return ui.writeDone(markedTask);
