+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,9 +1,9 @@
 package duke.commands;
 
 import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 
 /**
  * Command sub-type to define exiting the program
@@ -22,7 +22,8 @@ public class ExitCommand extends Command {
     }
 
     @Override
-    public String runNew(TaskList taskList, Storage storage, Ui ui) throws DukeException {
+    public String runGUI(TaskList taskList, Storage storage, Ui ui) throws DukeException {
+        System.exit(0);
         return ui.exit();
     }
 }
