+++ src/main/java/duke/util/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.util;
 
 import duke.tasks.Task;
