+++ src/main/java/duke/view/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.view;
 
 import java.io.IOException;
 import java.util.Collections;
